+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - N/A
 
+## [0.1.0] - 2024-05-15
+
+### Added
+
+- Added ConvertFrom-Jwt function
+- Added ConvertFrom-Jwt tests
+
 ## [0.0.9] - 2024-05-10
 
 ### Added